% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{europe_100km}
\alias{europe_100km}
\title{100-km grid of Europe}
\format{
A \code{sf} polygon.
\describe{
  \item{CellCode}{code for individual cells}
  \item{EofOrigin}{distance East of the data origin(centroid) in m}
  \item{NfOrigin}{distance North of the data origin(centroid) in m}
}
}
\source{
\url{https://www.eea.europa.eu/data-and-maps/data/eea-reference-grids-2}
}
\usage{
europe_100km
}
\description{
A grid of Europe at 100-km resolution in EPSG 3035. Based on FAO GAUL level 0 boundary polygons.
}
\author{
European Environment Agency (EEA)
}
\keyword{datasets}
