% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{europe_countries}
\alias{europe_countries}
\title{Country administrative boundaries}
\format{
\code{sf} polygons.
}
\source{
\url{https://www.naturalearthdata.com/downloads/}
}
\usage{
europe_countries
}
\description{
Natural Earth Data for countries within the European Bounding Box. Obtained through the \code{rnaturalearth} package at \code{scale = 50}. Reprojected to \code{ETRS89-extended / LAEA Europe} and clipped to a bounding box around Europe.
}
\author{
Natural Earth
}
\keyword{datasets}
