% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{europe_countries_midres}
\alias{europe_countries_midres}
\title{European country administrative boundaries}
\format{
\code{sf} polygons.
}
\source{
\url{https://www.naturalearthdata.com/downloads/}
}
\usage{
europe_countries_midres
}
\description{
Natural Earth data at \code{scale = 50} for countries within a European Bounding Box. Obtained through the \code{rnaturalearth} package at \code{scale = 50}. Reprojected to \code{ETRS89-extended / LAEA Europe} and clipped to a bounding box around Europe using the internal \code{N2000:::get_EuropeanCountries()} function.
}
\author{
Natural Earth
}
\keyword{datasets}
