% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{get_EuropeanCountries}
\alias{get_EuropeanCountries}
\title{Get administrative boundaries for European countries}
\usage{
get_EuropeanCountries(
  scale = 110,
  bbox = europe_100km,
  crs = sf::st_crs(europe_100km),
  ...
)
}
\arguments{
\item{scale}{(optional) The \code{numeric} scale to pass to \code{rnaturalearth::ne_countries()}. Acceptable values are \code{10} (highres), \code{50} (midres), \code{110} (lowres). Defaults to \code{110}.}

\item{bbox}{(optional) A \code{sf} object from which a bounding box can be retrieved to clip the countries boundaries. Defaults to the bounding box of the \code{europe_100km} grid included in the \code{rN2000} package.}

\item{crs}{(optional) A \code{crs}, \code{sf}, or \code{numeric} object giving a CRS to reproject the data to. If \code{crs} is numeric, this should be an EPSG code. Defaults to \code{ETRS89-extended / LAEA Europe}.}

\item{...}{other parameters passed to \code{rnaturalearth::ne_countries()}}
}
\value{
A \code{sf} object with the administrative boundaries of the relevant European countries.
}
\description{
Uses the \code{rnaturalearth} package to retrieve global administrative boundaries, which are reprojected, and then clipped to a bounding box around Europe.
}
\examples{
\dontrun{
get_EuropeanCountries()
}
}
\author{
Matt Lewis
}
\keyword{internal}
