% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Inverse of in call for convenience
Calculates the set of entries not present in the second vector}
\usage{
a \%notin\% b
}
\arguments{
\item{a}{First vector}

\item{b}{Second vector}
}
\description{
Inverse of in call for convenience
Calculates the set of entries not present in the second vector
}
\author{
Martin Jung
}
\keyword{internal}
