% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myLog}
\alias{myLog}
\title{Custom logging function for scripts}
\usage{
myLog(title = "[Processing] ", ...)
}
\arguments{
\item{title}{The title in the log output}

\item{...}{Any additional outputs or words for display}
}
\description{
Custom logging function for scripts
}
\author{
Martin Jung
}
\keyword{internal}
