% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_filterSpecies.R
\name{n2000_definitions}
\alias{n2000_definitions}
\title{#' Function to check a species name against a list of supplied names
#' Crosscheck against taxize reference system.
#' The output will be saved internally as new object and thus this function does not need to be rerun
#'
#' @param species_lists
#' @keywords internal}
\usage{
n2000_definitions(fname)
}
\arguments{
\item{fname}{File name pointing to the Natura2000_end2019_dataset_definitions.xls file}
}
\value{
Saves a list of definitions from the sheet in 'data'
}
\description{

}
\details{

}
\author{
Martin Jung
}
\keyword{internal}
