% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_filter.R
\name{n2000_filterSpecies}
\alias{n2000_filterSpecies}
\title{Create a species subset from the Natura 2000 data}
\usage{
n2000_filterSpecies(x, name, option, check.name = F)
}
\arguments{
\item{x}{The \code{character} file path to the Natura 2000 geopackage}

\item{name}{The \code{character} latin name of a focal species}

\item{option}{The \code{character} type of data to return. Options are \code{"PO"} ('presence only' - sites where the species is present), \code{"PA"} ('presence absence' - sites where the species is present or absent), or \code{"abundance"}.}

\item{check.name}{\code{logical} Check whether the supplied name exists in the Natura 2000 data? Defaults to \code{FALSE}.}
}
\value{
A \code{sf} object with the Natura 2000 sites containing this species
}
\description{
Create a species subset from the Natura 2000 data
}
\examples{
\dontrun{
n2000_filterSpecies('/filepath/Natura2000_end2019.gpkg', name = 'Asio otus')
}
}
\author{
Martin Jung

Matt Lewis
}
