% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_format.R
\name{n2000_format_definitions}
\alias{n2000_format_definitions}
\title{Load and reformat Natura 2000 definitions to a list.}
\usage{
n2000_format_definitions(x)
}
\arguments{
\item{x}{A [`character`] file path to the Natura 2000 data definitions file.}
}
\description{
Load and reformat Natura 2000 definitions to a list.
}
\author{
Matt Lewis

Martin Jung
}
