% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_format.R
\name{n2000_format_species}
\alias{n2000_format_species}
\title{Checks and attempts to fix erroneous species names from Natura 2000 data}
\usage{
n2000_format_species(x, overwrite = T, verbose = F, ...)
}
\arguments{
\item{x}{A \code{character} directory to the Natura 2000 csv files (i.e. containing 'Natura2000_end2019_HABITATS.csv'). Or a \code{character} file path to the Natura 2000 geopackage file ('Natura2000_end2019.gpkg').}

\item{overwrite}{A \code{logical} input - should the Natura 2000 data be overwritten with the fixed names? If no, only an internal R object is returned. Defaults to \code{TRUE}.}

\item{verbose}{A \code{logical} input. Be chatty? Defaults to \code{FALSE}.}

\item{...}{Other parameters passed on to \code{taxize::tax_name()} or \code{taxize::synonyms()}}
}
\description{
Uses the \code{taxize} package to look-up and (hopefully) fix incorrect species names.
}
\author{
Matt Lewis
}
