% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_get.R
\name{n2000_getCSV}
\alias{n2000_getCSV}
\title{Set or download the N2000 database in csv format to a given folder}
\usage{
n2000_getCSV(
  folder,
  year = "2019",
  force = FALSE,
 
    nlink = "https://www.eea.europa.eu/data-and-maps/data/natura-11/natura-2000-tabular-data-12-tables/natura-2000-comma-separated-values-files/at_download/file"
)
}
\arguments{
\item{folder}{A [`character`] directory path to download the Natura 2000 data set to. Automatic checks to see if the data already exist in the folder (can be overridden by specifying `force = T`).}

\item{year}{(optional). The [`character`] year version of the Natura 2000 data set to check if it already exists. Defaults to the most recently available version (2019).}

\item{force}{(optional). [`logical`] should a new download be forced even if data already exists in the directory? Defaults to `FALSE`.}

\item{nlink}{(normally not altered). The [`character`] direct download link for Natura 2000.}
}
\description{
Set or download the N2000 database in csv format to a given folder
}
\examples{
\dontrun{
n2000_getCSV(folder = 'myfolder', year = "2019")
}
}
\author{
Martin Jung

Matt Lewis
}
