% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_get.R
\name{n2000_getGPGK}
\alias{n2000_getGPGK}
\title{Set or download the N2000 database in geopackage format to a given folder}
\usage{
n2000_getGPGK(
  folder,
  year = "2019",
  force = FALSE,
  nlink = "https://cmshare.eea.europa.eu/s/GkqdcbbsYmmBSEQ/download"
)
}
\arguments{
\item{folder}{A [`character`] directory path to download the Natura 2000 data set to. Automatic checks to see if the data already exist in the folder (can be overridden by specifying `force = T`).}

\item{year}{(optional). The [`character`] year version of the Natura 2000 data set to check if it already exists. Defaults to the most recently available version (2019).}

\item{force}{(optional). [`logical`] should a new download be forced even if data already exists in the directory? Defaults to `FALSE`.}

\item{nlink}{(normally not altered). The [`character`] direct download link for Natura 2000.}
}
\description{
Set or download the N2000 database in geopackage format to a given folder
}
\examples{
\dontrun{
n2000_getGPKG(folder = 'myfolder', year = "2019")
}
}
\author{
Martin Jung

Matt Lewis
}
