% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_filterSpecies.R
\name{n2000_habitat}
\alias{n2000_habitat}
\title{Create a habitat subset from the Natura 2000 data}
\usage{
n2000_habitat(x, name)
}
\arguments{
\item{x}{The \code{character} file path to the Natura 2000 geopackage}

\item{name}{The \code{character} name of the habitat}
}
\value{
A \code{sf} object with the Natura 2000 sites containing this habitat
}
\description{
Create a habitat subset from the Natura 2000 data
}
\examples{
\dontrun{
n2000_habitat('/filepath/Natura2000_end2019.gpkg', name = 'Mountain hay meadows')
}
}
\author{
Martin Jung

Matt Lewis
}
