% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_intersect.R
\name{n2000_intersect}
\alias{n2000_intersect}
\title{Extract information from Natura 2000 sites within a target area of interest by means of spatial intersection.}
\usage{
n2000_intersect(n2000_file, target_areas, byid = T, enclosed = F)
}
\arguments{
\item{n2000_file}{[`character`] or [`object`] The file path to the Natura2000 geopackage. Alternatively the name of the loaded geopackage in the environment.}

\item{target_areas}{[`character`] or [`object`] The file path to the spatial polygon(s) for the area of interest. Must be readable by \code{sf::st_read()}. Alternatively the name of the loaded spatial data in the environment.}

\item{byid}{[`logical`]. Should data be extracted for each target area (TRUE) or should they be merged first (FALSE)? Defaults to TRUE.}

\item{enclosed}{[`logical`]. Should Natura 2000 sites only be included if they are fully enclosed by the target area? Defaults to FALSE.}
}
\description{
Extract Natura 2000 site information which overlap specified spatial locations.
}
\details{
If the CRS of the \code{target_areas} does not match that of the Natura 2000 data, the \code{target_areas} are first reprojected.
}
\author{
Matt Lewis
}
