% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_get.R
\name{n2000_showdefinitions}
\alias{n2000_showdefinitions}
\title{Show N2000 database definitions}
\usage{
n2000_showdefinitions(table = NULL)
}
\arguments{
\item{table}{Which table of the definitions (optional)}
}
\description{
Show N2000 database definitions
}
\examples{
\dontrun{
def_list <- n2000_getdefinitions()
def_list$HABITATS
}
}
\author{
Martin Jung
}
