% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_spatial.R
\name{n2000_SpatialSpecies}
\alias{n2000_SpatialSpecies}
\title{Prepares a target species multipolygon feature for modelling
using}
\usage{
n2000_SpatialSpecies(
  species,
  metric = "pa",
  abundance_bound = "UPPERBOUND",
  what = "polygon",
  ref_grid = "1km"
)
}
\arguments{
\item{species}{A data.frame obtained via the n2000_filterSpecies function}

\item{metric}{How is biodiversity encoded in the output. As presence only ('po'), presence absence ('pa') or a population estimate ('abundance')}

\item{abundance_bound}{Which estimate to use for metric option 'abundance'. Default is 'UPPERBOUND'}

\item{what}{What should be created. Possible options ('polygon','raster')}

\item{ref_grid}{Which European reference grid to use. Default is '1km'}
}
\value{
A spatial object of the respective species. The
}
\description{
Prepares a target species multipolygon feature for modelling
using
}
\examples{
\dontrun{
species <- n2000_species('/media/martin/data/EuropeanBiodiversityData/N2000/Natura2000_end2019.gpkg',
                    sname = 'Asio otus')
o <- n2000_SpatialSpecies(species, 'Asio_otus', metric = 'pa', what = 'raster' )
plot(o)
}
}
\author{
Martin Jung
}
