% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_filterSpecies.R
\name{n2000_species}
\alias{n2000_species}
\title{Create a species subset from the Natura2000 dataset}
\usage{
n2000_species(n2000_file, sname)
}
\arguments{
\item{n2000_file}{The file to the Natura2000 geopackage}

\item{sname}{The latin name of a focal species}
}
\value{
A sf object with the Natura 2000 sites containing this species
}
\description{
Create a species subset from the Natura2000 dataset
}
\examples{
\dontrun{
n2000_species('/media/martin/data/EuropeanBiodiversityData/N2000/Natura2000_end2019.gpkg', sname = 'Asio otus')
}
}
\author{
Martin Jung
}
