% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_format.R
\name{n2000_unique_sp}
\alias{n2000_unique_sp}
\title{Get list of unique species names from Natura 2000 checklists}
\usage{
n2000_unique_sp(x)
}
\arguments{
\item{x}{A [`character`] directory to the Natura 2000 csv files (i.e. containing 'Natura2000_end2019_SPECIES.csv'). Or a [`character`] file path to the Natura 2000 geopackage file ('Natura2000_end2019.gpkg').}
}
\value{
A \code{vector} of species binomial names.
}
\description{
Get list of unique species names from Natura 2000 checklists
}
\details{
Works with either geopackage (.gpkg) or csv (.csv) file extensions of downloaded data.

If you haven't downloaded the data yet, see `rN2000::n2000_getCSV()` or `rN2000::n2000_getGPKG()`
}
\author{
Matt Lewis
}
