% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_spatialize.R
\name{prep_EUgrid}
\alias{prep_EUgrid}
\title{Downloads the EU reference grid and saves it as data.object
The European-wide reference grid is only at 10km resolution}
\usage{
prep_EUgrid(url)
}
\arguments{
\item{url}{The url to the download site}
}
\value{
Saves as data object the reference grid at varying resolutions
}
\description{
Downloads the EU reference grid and saves it as data.object
The European-wide reference grid is only at 10km resolution
}
\examples{
\dontrun{
prep_EUgrid('https://www.eea.europa.eu/data-and-maps/data/eea-reference-grids-2/gis-files/europe-10-km-100-km/at_download/file')
}
}
\author{
Martin Jung
}
\keyword{internal}
