% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeGeoTiff}
\alias{writeGeoTiff}
\title{Write raster files as GeoTiffs with useful compression options}
\usage{
writeGeoTiff(x, file, datatype = "FLT4S")
}
\arguments{
\item{x}{A \code{raster} object.}

\item{file}{A \code{character} file path and file name (e.g. \code{'filepath/myraster.tif'}).}

\item{datatype}{A \code{character} input for data type. Defaults to \code{'FLT4S'}, but \code{'LOG1S'} is useful for binary rasters. See \code{?raster::dataType()}}

\item{...}{Other parameters passed on to \code{raster::writeRaster()}}
}
\description{
Write raster files as GeoTiffs with useful compression options
}
\details{
Defaults to using the following parameters:
\code{NAflag = -9999}; \code{options = c("COMPRESS=DEFLATE","PREDICTOR=2", "ZLEVEL=9")}; \code{format = "GTiff"}; \code{overwrite = T}
}
\author{
Martin Jung

Matt Lewis
}
\keyword{internal}
