% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_get.R
\name{n2000_definitions}
\alias{n2000_definitions}
\title{Internal unction to prepare the definitions file for further use}
\usage{
n2000_definitions(fname)
}
\arguments{
\item{fname}{File name pointing to the Natura2000_end2019_dataset_definitions.xls file}
}
\value{
Saves a list of definitions from the sheet in 'data'
}
\description{
Internal unction to prepare the definitions file for further use
}
\author{
Martin Jung
}
\keyword{internal}
