% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_get.R
\name{n2000_getCSV}
\alias{n2000_getCSV}
\title{Set or download the N2000 database in csv format to a given folder}
\usage{
n2000_getCSV(
  folder,
  year = "2019",
 
    n_link = "https://www.eea.europa.eu/data-and-maps/data/natura-11/natura-2000-tabular-data-12-tables/natura-2000-comma-separated-values-files/at_download/file"
)
}
\arguments{
\item{folder}{A folder with the Natura 2000 dataset}

\item{year}{The latest version of the Natura 2000 dataset}

\item{nlink}{This should normally not be altered. Direct download link for Natura 2000}
}
\description{
Set or download the N2000 database in csv format to a given folder
}
\examples{
\dontrun{
n2000_getCSV(folder = 'myfolder', year = "2019")
}
}
\author{
Martin Jung
}
