% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_get.R
\name{n2000_getGPGK}
\alias{n2000_getGPGK}
\title{Set or download the N2000 database in geopackage format to a given folder}
\usage{
n2000_getGPGK(
  folder,
  year = "2019",
  n_link = "https://cmshare.eea.europa.eu/s/GkqdcbbsYmmBSEQ/download"
)
}
\arguments{
\item{folder}{A folder with the Natura 2000 dataset}

\item{year}{The latest version of the Natura 2000 dataset}

\item{nlink}{This should normally not be altered. Direct download link for Natura 2000}
}
\description{
Set or download the N2000 database in geopackage format to a given folder
}
\examples{
\dontrun{
n2000_getGPKG(folder = 'myfolder', year = "2019")
}
}
\author{
Martin Jung
}
