% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_filterSpecies.R
\name{n2000_habitat}
\alias{n2000_habitat}
\title{Create a habitat subset from the Natura2000 dataset}
\usage{
n2000_habitat(n2000_file, hname)
}
\arguments{
\item{n2000_file}{The file to the Natura2000 geopackage}

\item{hname}{The name of the habitat}
}
\value{
A sf object with the Natura 2000 sites containing this habitat
}
\description{
Create a habitat subset from the Natura2000 dataset
}
\examples{
\dontrun{
n2000_habitat('/media/martin/data/EuropeanBiodiversityData/N2000/Natura2000_end2019.gpkg', hname = 'Mountain hay meadows')
}
}
\author{
Martin Jung
}
