% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_mapwms.R
\name{n2000_mapwms}
\alias{n2000_mapwms}
\title{Maps outs the Natura 2000 sites using the commissions WMS services}
\usage{
n2000_mapwms(
  title = "N2000",
  clng = 16.3588059165812,
  clat = 48.0618047198572,
  czoom = 12
)
}
\arguments{
\item{title}{The title of the WMS layer to be rendered}
}
\description{
This function requires the {leaflet} package to be installed and a functioning internet connection
}
\examples{
\dontrun{
n2000_mapwms(folder = 'myfolder', year = "2019")
}
}
\author{
Martin Jung
}
