% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_mapwms.R
\name{n2000_mapwms}
\alias{n2000_mapwms}
\title{Map Natura 2000 sites using WMS services}
\usage{
n2000_mapwms(
  title = "N2000",
  clng = 16.3588059165812,
  clat = 48.0618047198572,
  czoom = 12
)
}
\arguments{
\item{title}{The [`character`] title of the WMS layer to be rendered. Options are: 'N2000', 'CDDA', and 'N2000_centroids'. Defaults to 'N2000'. See details.}

\item{clng}{The [`numeric`] centre-point longitude coordinate of the field of view. See details.}

\item{clat}{The [`numeric`] centre-point latitude coordinate of the field of view. See details.}

\item{czoom}{The [`numeric`] zoom level to use. Defaults to 12 (see details).}
}
\description{
N.b. This function requires the {`leaflet`} package to be installed and a functioning internet connection.
}
\details{
`title` has three options: 'N2000' to render Natura 2000 sites, 'CDDA' to render sites listed on the Common Database on Designated Areas (i.e. nationally designated areas) or 'N2000_centroids' to render only the centroids of the Natura 2000 sites.

The default field of view for the returned image focuses on the Schlosspark in Laxenburg, Austria. This is arbitrary, and does not by any means reflect any biases on the parts of the authors.
}
\examples{
\dontrun{
n2000_mapwms()
}
}
\author{
Martin Jung

Matt Lewis
}
