% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_plot.R
\name{n2000_plot}
\alias{n2000_plot}
\title{Plots retrieved Natura 2000 with optional grid, and country borders}
\usage{
n2000_plot(x, borders = F, grid = F)
}
\arguments{
\item{x}{A \code{sf} object to plot.}

\item{borders}{A \code{logical} input - should country borders be plotted too? Defaults to \code{FALSE}.}

\item{grid}{A \code{logical} input. Should a grid be added too? Defaults to \code{FALSE}.}
}
\description{
Wrapper around ggplot for convenience. Requires `ggplot2` to be installed.
}
\author{
Matt Lewis
}
