% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2000_filterSpecies.R
\name{n2000_species}
\alias{n2000_species}
\title{Create a species subset from the Natura 2000 data}
\usage{
n2000_species(x, name)
}
\arguments{
\item{x}{The \code{character} file path to the Natura 2000 geopackage}

\item{name}{The \code{character} latin name of a focal species}
}
\value{
A \code{sf} object with the Natura 2000 sites containing this species
}
\description{
Create a species subset from the Natura 2000 data
}
\examples{
\dontrun{
n2000_species('/filepath/Natura2000_end2019.gpkg', name = 'Asio otus')
}
}
\author{
Martin Jung

Matt Lewis
}
